renamed evapotranspiration, wrote test for evaporate()
